Separate auth provider from App.tsx
