+++ src/App.tsx
@@ -1,8 +1,7 @@
-import { useState } from 'react';
+import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import AppBar from './features/AppBar/AppBar';
 import AppFooter from './features/AppFooter/AppFooter';
-import { AuthContext } from './shared/auth-context';
 import { PossibleRoutes } from './utils/constants';
 import MainPage from './pages/MainPage/MainPage';
 import DashboardPage from './pages/DashboardPage/DashboardPage';
@@ -10,15 +9,9 @@ import { ThemeProvider } from '@mui/material/styles';
 import theme from './theme/theme';
 import { initializeApp } from 'firebase/app';
 import firebaseConfig from './configs/firebase.configs';
-import { 
-  getAuth, 
-  signOut, 
-  setPersistence, 
-  browserLocalPersistence, 
-  onAuthStateChanged
-} from 'firebase/auth';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { ReactQueryDevtools } from 'react-query/devtools';
+import AuthProvider from './App.authProvider';
 
 export const endpoint = 'http://localhost:9000/graphql';
 
@@ -34,28 +27,6 @@ export const queryClient = new QueryClient({
 initializeApp(firebaseConfig);
 
 const App = () => {
-  const auth = getAuth();
-  setPersistence(auth, browserLocalPersistence);
-  const [userId, setUserId] = useState<string | undefined>(undefined);
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-  const [displayName, setDisplayName] = useState<string | null>('');
-  const [photoURL, setPhotoURL] = useState<string | null>('');
-  const [expectedDueDate, setExpectedDueDate] = useState<string | null>(null)
-  
-  onAuthStateChanged(auth, (user) => {
-    if (user) {
-      setUserId(user.uid);
-      setIsLoggedIn(true);
-      setDisplayName(user.displayName);
-      setPhotoURL(user.photoURL);
-    } else {
-      setUserId(undefined);
-      setIsLoggedIn(false);
-      setDisplayName('');
-      setPhotoURL('');
-      signOut(auth);
-    }
-  })
 
   const routes = (
     <Routes>
@@ -67,15 +38,7 @@ const App = () => {
   
   return (
     <QueryClientProvider client={queryClient}>
-      <AuthContext.Provider 
-        value={{
-          isLoggedIn: isLoggedIn,
-          id: userId,
-          displayName: displayName,
-          photoURL: photoURL,
-          expectedDueDate: expectedDueDate
-        }}
-      >
+      <AuthProvider>
         <ThemeProvider theme={theme}>
           <Router>
             <ReactQueryDevtools initialIsOpen={false} />
@@ -86,7 +49,7 @@ const App = () => {
             <AppFooter />
           </Router>
         </ThemeProvider>
-      </AuthContext.Provider>
+      </AuthProvider>
     </QueryClientProvider>
   );
 };
